Move write_state_file_atomically to part07 to stay under file-size cap
The spawn_blocking rewrite pushed app_state_impl_parts/part01.rs to 2013
lines, over the 2000-line Lint Backend hard cap. All app_state_impl_parts
share one module scope via include!, and free fns are order-independent,
so relocate the helper next to the other state utility fns (sha256_hex,
constant_time_str_eq) in part07.rs. part01 drops to 1980; no behavior
change.
